introduce Camera Proxy
With File Shuttle, users can choose cross-profile files to be used in apps that support Documents UI and SAF. But it turns out that we can do better - even in some apps that do not support SAF natively, they may however support invoking external Camera apps to take photos for them. Shelter can present as a fake Camera app, returning user-chosen images as the "photos". In this way, it is much more convenient to transfer pictrues in and out of Shelter, at least for me it makes me feel way better when I want to post screenshots etc. from apps inside Shelter.