AppListFragment: do not hide launch option if launch intent is not found
This consideration was added to avoid accidentally launching system apps
that are normally hidden from the list but can be shown by the new
option. However, "frozen" apps also have no launch intents, but they
should be able to be "unfrozen and launched". Thus, we just ignore this
check and show a toast notifying user about the failure if an app
without GUI gets launched.