ShelterDeviceAdminReceiver: use notification to start activity
On Android 10, it is no longer possible to start an activity without
being visible in the first place. However, Shelter depends on starting
the activity from the DAR to initialize the provisioned work profile. We
switch to a notification with the highest importance to let the user
manually bring up the DummyActivity for finalization.

This may also fix bugs on devices running EMUI / MIUI that were
previously not working due to similar reasons.