SettingsFragment: disable FileShuttle on Q for now
FileShuttle needs to be totally refactored to be able to support Q. For
now, let's first make sure basic features are working and just disable
FileShuttle for a while.

Users upgraded from lower versions of Android to Q would not be affected
because Shelter already has permissions to read and write external
storage.