Use ACTION_PROVISIONING_SUCCESSFUL instead of ACTION_PROFILE_PROVISIONING_COMPLETE
On Android 8.0+, the activity intent ACTION_PROVISIONING_SUCCESSFUL is
also sent out before the broadcast
ACTION_PROFILE_PROVISIONING_COMPLETE. We can make of use this to
directly bring up an activity for finalization, skipping over our
original notification-based switch-er-oo from the broadcast receiver.

This is also necessary on platform release T since now we cannot send
notifications by default. We cannot request for the permission either
since that will require a visible activity too. Using the activity
intent works around all of this and makes Shelter work again on T. (And
we won't have to deal with the Do-Not-Disturb nonsense anymore)